Moved tests to typescript
