+++ package.json
@@ -73,6 +73,7 @@
         "uuid": "~3.4.0"
     },
     "devDependencies": {
+        "@types/mocha": "^8.0.3",
         "@types/node": "^14.0.27",
         "@typescript-eslint/eslint-plugin": "^3.9.0",
         "@typescript-eslint/parser": "^3.9.0",
