+++ src/index.ts
@@ -1,18 +1,7 @@
 "use strict";
 
-export { KStream } from "./lib/dsl/KStream";
-export { KTable } from "./lib/dsl/KTable";
+export { KStream, KTable } from "./lib/dsl";
 export { KafkaFactory } from "./lib/KafkaFactory";
 export { KafkaStreams } from "./lib/KafkaStreams";
-export { KStorage } from "./lib/KStorage.js";
+export { KStorage } from "./lib/KStorage";
 export { KafkaClient } from "./lib/client/KafkaClient";
-
-// module.exports = {
-//   default: KafkaStreams,
-//   KStream,
-//   KTable,
-//   KafkaFactory,
-//   KafkaStreams,
-//   KStorage,
-//   KafkaClient
-// };
