+++ src/lib/actions/Min.ts
@@ -1,4 +1,4 @@
-import Promise from "bluebird";
+import { Promise } from "bluebird";
 
 /**
  * used grab the lowest value of
